refactor: 금액에 대한 검증과 수량을 계산하는 Money 클래스 생성 (step2 리뷰)
